Exercise 05 - Portfolio
In this exercise you will continue building the AlgoTrader application
by creating a better
design to handle your stocks data.
In order to manage your stocks in one place you will create a Portfolio
​class (Portfolio.java)
that has an array of your stocks.
In order to activate actions of portfolio you will add a
PortfolioManager​class, and in order to
present the portfolio data in web page you will create a new servlet:
PorfolioServlet​